refactor: block 관련 DTO를 relation로 이동
