+++ app-main/src/main/java/net/causw/app/main/domain/user/account/api/v2/mapper/BlockDtoMapper.java
@@ -3,8 +3,8 @@ package net.causw.app.main.domain.user.account.api.v2.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
-import net.causw.app.main.domain.community.block.service.dto.ChildCommentBlockCreateCommand;
-import net.causw.app.main.domain.community.block.service.dto.CommentBlockCreateCommand;
+import net.causw.app.main.domain.user.relation.service.v2.dto.ChildCommentBlockCreateCommand;
+import net.causw.app.main.domain.user.relation.service.v2.dto.CommentBlockCreateCommand;
 import net.causw.app.main.domain.user.account.api.v2.dto.response.BlockResponseDto;
 import net.causw.app.main.domain.user.account.entity.user.User;
 import net.causw.app.main.domain.user.relation.service.v2.dto.BlockCreateCommand;
