+++ app-main/src/main/java/net/causw/app/main/domain/user/relation/service/v2/BlockService.java
@@ -3,8 +3,8 @@ package net.causw.app.main.domain.user.relation.service.v2;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import net.causw.app.main.domain.community.block.service.dto.ChildCommentBlockCreateCommand;
-import net.causw.app.main.domain.community.block.service.dto.CommentBlockCreateCommand;
+import net.causw.app.main.domain.user.relation.service.v2.dto.ChildCommentBlockCreateCommand;
+import net.causw.app.main.domain.user.relation.service.v2.dto.CommentBlockCreateCommand;
 import net.causw.app.main.domain.community.comment.entity.ChildComment;
 import net.causw.app.main.domain.community.comment.entity.Comment;
 import net.causw.app.main.domain.community.comment.service.implementation.ChildCommentReader;
