+++ app-main/src/main/java/net/causw/app/main/domain/user/relation/service/v2/dto/CommentBlockCreateCommand.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.community.block.service.dto;
+package net.causw.app.main.domain.user.relation.service.v2.dto;
 
 import net.causw.app.main.domain.user.account.entity.user.User;
